--- a/doc/Makefile
+++ b/doc/Makefile
@@ -49,8 +49,9 @@ help:
 clean:
 	-rm -rf $(BUILDDIR)/*
 	-rm -rf auto_examples/
-	-rm -rf generated/*
-	-rm -rf modules/generated/*
+	-rm -rf generated/
+	-rm -rf reference/generated/
+	-rm -rf modules/generated/
 
 html:
 	# These two lines make the build a bit more lengthy, and the