--- a/pydicom/filebase.py
+++ b/pydicom/filebase.py
@@ -3,7 +3,7 @@
 
 from __future__ import absolute_import
 
-from pydicom.tag import Tag, BaseTag
+from pydicom.tag import Tag
 from struct import (unpack, pack)
 
 from io import BytesIO
@@ -38,7 +38,7 @@ class DicomIO(object):
     def write_tag(self, tag):
         """Write a dicom tag (two unsigned shorts) to the file."""
         # make sure is an instance of class, not just a tuple or int
-        if not isinstance(tag, BaseTag):
+        if not isinstance(tag, Tag):
             tag = Tag(tag)
         self.write_US(tag.group)
         self.write_US(tag.element)