… doc/old/base_element.rst
… doc/whatsnew/v1.4.0.rst
… pydicom/values.py