#### What does this implement/fix? Explain your changes.
Unifies `tag.Tag()` with `tag.BaseTag`. Essentially `BaseTag.__new__()` is overridden so that it accepts the same arguments as `Tag()` and also allows the use of a `fast` keyword parameter so that `TupleTag()` can still create fast objects. In addition, `tag.Tag()` is removed and `BaseTag` renamed to`Tag` with an alias to `BaseTag` to allow backwards compatibility.

The idea is to simplify access and creation of elements tags (see the [tags section](https://1468-14006067-gh.circle-artifacts.com/0/home/circleci/project/doc/_build/html/old/base_element.html#tag) in the documentation to see what I mean)

I also added a deprecation for `BaseTag` with removal in v1.5