@pep8speaks 2019-07-31
Hello @scaramallion! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:













* In the file [`pydicom/tag.py`](https://github.com/pydicom/pydicom/blob/2bec27b40c9c8352c8e0d6caa111ec619c7cba2c/pydicom/tag.py):

> [Line 118:80](https://github.com/pydicom/pydicom/blob/2bec27b40c9c8352c8e0d6caa111ec619c7cba2c/pydicom/tag.py#L118): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (81 > 79 characters)
> [Line 146:80](https://github.com/pydicom/pydicom/blob/2bec27b40c9c8352c8e0d6caa111ec619c7cba2c/pydicom/tag.py#L146): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (82 > 79 characters)