@scaramallion 2019-08-02
Its a little bit slower than before for some reason, not sure why. It might be that an `int.__new__` override isn't as fast as without? I was hoping it would end up faster but I don't think its too bad.

Time per 100,000 runs

Current master
```
TimeTag.time_keyword                                          826.36ms
TimeTag.time_list_int                                         307.58ms
TimeTag.time_list_str                                         346.48ms
TimeTag.time_single_int                                       182.43ms
TimeTag.time_single_str                                       222.79ms
TimeTag.time_single_tag                                        42.59ms
TimeTag.time_tuple_int                                        274.40ms
TimeTag.time_tuple_str                                        346.95ms
```

This PR
```
TimeTag.time_keyword                                          882.12ms
TimeTag.time_list_int                                         333.16ms
TimeTag.time_list_str                                         391.23ms
TimeTag.time_single_int                                       227.56ms
TimeTag.time_single_str                                       268.32ms
TimeTag.time_single_tag                                        61.82ms
TimeTag.time_tuple_int                                        305.96ms
TimeTag.time_tuple_str                                        381.30ms
```