@scaramallion 2019-08-02
Preview of the changed documentation page [here](https://1483-14006067-gh.circle-artifacts.com/0/home/circleci/project/doc/_build/html/old/base_element.html) and the `Tag` API page [here](https://1483-14006067-gh.circle-artifacts.com/0/home/circleci/project/doc/_build/html/reference/generated/pydicom.tag.Tag.html#pydicom.tag.Tag)