@darcymason 2019-08-02
@scaramallion, can you remind me how to get at the timing results for this?  Are they available online? If not, do you have a config file for asv to get me started locally?


> The idea is to simplify access and creation of elements tags (see the [tags section](https://1468-14006067-gh.circle-artifacts.com/0/home/circleci/project/doc/_build/html/old/base_element.html#tag) in the documentation to see what I mean)

Can you clarify a little more on this?  Is it just that the Tag factory returning BaseTag seems a little confusing?