@scaramallion 2019-08-03
> Can you clarify a little more on this? Is it just that the Tag factory returning BaseTag seems a little confusing?

Yeah pretty much, we have `Tag` (which isn't a tag) that returns `BaseTag` (which is). It seems simpler and less confusing to just have a single class given that's effectively what we're doing anyway.

The other alternative I considered was to alias `Tag()` as `create_tag()` and alias `BaseTag` as `Tag` but that would break backwards compatibility and given that both `Tag` and `BaseTag` have been around forever and are probably reasonably commonly used (pynetdicom uses `Tag()`,  for example) I thought this would be better. 

> @scaramallion, can you remind me how to get at the timing results for this? Are they available online? If not, do you have a config file for asv to get me started locally?



[Here](https://github.com/pydicom/pydicom/files/3463208/asv.conf.json.txt) is my conf file, just remove the `.txt` extension and run `asv dev -b Tag` in the pydicom root directory.

I just run them on whatever computer I'm working on. As long as the before and after are on the same PC then they should be indicative of the changes. Results do vary between different PCs, obviously.