@darcymason 2019-08-04
> Yeah pretty much, we have `Tag` (which isn't a tag) that returns `BaseTag` (which is). It seems simpler and less confusing to just have a single class given that's effectively what we're doing anyway

Tag was originally a class and was changed to a factory function - that's why it has a capitalized name.  I'm trying to find some history for why that happened.  I suspect it was for speed, but perhaps our changes to checking isinstance before calling Tag has done more to improve that anyway.