@darcymason 2019-08-04
> I suspect it was for speed, but perhaps our changes to checking isinstance before calling Tag has done more to improve that anyway.

Looks like it was in 32ab5078, from Jan 2010, and is in one commit with lots of changes for speed, so hard to tell how much effect.  Hell of a long time ago.

I'll look a bit at timing and think about this a bit more.  The timing from your table was mostly ~10% slower.  Single tag stands out at ~50% more time, but is now mostly avoided with the `isinstance` checks in pydicom.  Single int was ~25% more,  I'd be interested to know how that translates into reading files.   If only a few % overall, then is probably worth it to make it a class again for clarity.

On a related note, I've realized for a long time that the tag handling is messy (as all the added `isinstance` checks showed).  It is probably more beneficial to eventually clean up as many functions as possible to _know_ whether they have a Tag instance or not and avoid extra checks or conversions.