@scaramallion 2019-08-05
I think it could be refactored a bit to prioritise the single int case, that is what we're mostly concerned with. I'll play around and see what I can do.