@scaramallion 2019-08-05
Current master (I added benchmarks for the double int, double str case)

```
TimeTag.time_double_int                                        299.85ms
TimeTag.time_double_str                                        357.22ms
TimeTag.time_keyword                                           821.79ms
TimeTag.time_list_int                                          296.59ms
TimeTag.time_list_str                                          364.40ms
TimeTag.time_single_int                                        175.68ms
TimeTag.time_single_str                                        217.13ms
TimeTag.time_single_tag                                         42.10ms
TimeTag.time_tuple_int                                         287.36ms
TimeTag.time_tuple_str                                         353.32ms
```

Prioritising single int 
```
TimeTag.time_double_int                                        285.08ms
TimeTag.time_double_str                                        446.31ms
TimeTag.time_keyword                                           934.57ms
TimeTag.time_list_int                                          339.58ms
TimeTag.time_list_str                                          486.18ms
TimeTag.time_single_int                                        161.02ms
TimeTag.time_single_str                                        302.68ms
TimeTag.time_single_tag                                         61.21ms
TimeTag.time_tuple_int                                         324.95ms
TimeTag.time_tuple_str                                         465.98ms
```

The single int case is now faster than current master. Looking at the code base where `Tag` is used there are a few spots that could be optimised. I think that'd be better in a new PR.