@scaramallion 2019-08-05
Dropping support for the str/double str and list/tuple of str cases would help a bit too, I doubt they get used much. However, I think the keyword use case is far too useful to drop, but its pretty slow so we'll have to watch out that it doesn't make it into the code base more than required.